Move frame related kws to own lib component
