+++ src/SeleniumLibrary/__init__.py
@@ -25,6 +25,7 @@ from SeleniumLibrary.keywords import (AlertKeywords,
                                       CookieKeywords,
                                       ElementKeywords,
                                       FormElementKeywords,
+                                      FrameKeywords,
                                       JavaScriptKeywords,
                                       RunOnFailureKeywords,
                                       ScreenshotKeywords,
@@ -329,14 +330,15 @@ class SeleniumLibrary(DynamicCore):
         libraries = [
             AlertKeywords(self),
             BrowserManagementKeywords(self),
-            RunOnFailureKeywords(self),
-            ElementKeywords(self),
-            TableElementKeywords(self),
-            FormElementKeywords(self),
-            SelectElementKeywords(self),
-            JavaScriptKeywords(self),
             CookieKeywords(self),
+            ElementKeywords(self),
+            FormElementKeywords(self),
+            FrameKeywords(self),
+            JavaScriptKeywords(self),
+            RunOnFailureKeywords(self),
             ScreenshotKeywords(self),
+            SelectElementKeywords(self),
+            TableElementKeywords(self),
             WaitingKeywords(self)
         ]
         self._browsers = BrowserCache()
